modified 'ListingController.php', 'Listing.php', 'User.php', 'RouteServiceProvider.php', '2023_11_08_080600_create_listings_table.php', 'DatabaseSeeder.php', 'layout.blade.php', 'show.blade.php', and 'web.php' then added 'UserController.php', 'manage.blade.php', 'login.blade.php', and 'register.blade.php'

+++ app/Http/Controllers/ListingController.php
@@ -48,6 +48,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
         return redirect('/')->with('message', 'Listing created successfully!');
     }
@@ -59,27 +61,44 @@ class ListingController extends Controller
 
     // Update listing data
     public function update(Request $request, Listing $listing) {
-        $formFields = $request->validate([
-            'title' => 'required',
-            'company' => 'required',
-            'location' => 'required',
-            'website' => 'required',
-            'email' => ['required', 'email'],
-            'tags' => 'required',
-            'description' => 'required'
-        ]);
 
-        if($request->hasFile('logo')){
-            $formFields['logo'] = $request->file('logo')->store('logos', 'public');
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            return abort(403, 'Unauthorized action');
+        }else{
+            $formFields = $request->validate([
+                'title' => 'required',
+                'company' => 'required',
+                'location' => 'required',
+                'website' => 'required',
+                'email' => ['required', 'email'],
+                'tags' => 'required',
+                'description' => 'required'
+            ]);
+    
+            if($request->hasFile('logo')){
+                $formFields['logo'] = $request->file('logo')->store('logos', 'public');
+            }
+    
+            $listing->update($formFields);
+            return back()->with('message', 'Listing updated successfully!');
         }
-
-        $listing->update($formFields);
-        return back()->with('message', 'Listing updated successfully!');
     }
 
     // Destroy listing
     public function destroy(Listing $listing){
-        $listing->delete();
-        return redirect('/')->with('message', 'Listing deleted successfully');
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            return abort(403, 'Unauthorized action');
+        }else{
+            $listing->delete();
+            return redirect('/')->with('message', 'Listing deleted successfully');
+        }
+    }
+
+    // Manage listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
